requirements.txt
python-dotenv == 1.1.1
langchain-community == 0.4.1
langchain-text-splitters == 1.1.1
streamlit == 1.46.1
langchain-huggingface == 1.2.1
langchain-groq == 1.1.2
langchain-chroma == 1.1.0
unstructured==0.18.32




